refactor(profile): move the image reposition drag out of the card
The card renders five card formats and also ran a second, unrelated gesture:
dragging a cropped image inside its frame to choose what stays visible. That
is its own piece of state — a working position kept apart from the saved one,
the measured aspect ratio, and the overflow the drag is scaled against.

It moves to useProfileImageReposition, which takes the image URL and the saved
position and hands back the frame ref, the object-position style, and the
pointer handlers. The card drops from 674 lines to 588.
